+++ mols2grid/__init__.py
@@ -1,3 +1,4 @@
 from .molgrid import MolGrid
 from .dispatch import display, save
+from .select import selection, reset_selection, set_selection, del_selection
 from ._version import __version__